refactor(trace): extract computed view properties into View structs
